requirements.txt
tesseract-ocr
pytesseract
pillow
easyocr
requests
BeautifulSoup4